Refactor frame encoding
Frame.Encoder is now a protocol with a default implementation which does
frame encoding and decoding. Standard frames use this.

Frame.Flags and Frame.Payload are the protocols used for (d)ecoding the
flags and payload structs. The default implementation does nothing.

Frame now simply injects the Frame structure into the frame modules.